move unchecked iterator
