+++ sdk/src/lib.rs
@@ -27,9 +27,11 @@ use env::*;
 
 mod bigint;
 mod binary;
+mod iter;
 mod map;
 mod vec;
 pub use bigint::BigInt;
 pub use binary::{ArrayBinary, Binary, FixedLengthBinary, VariableLengthBinary};
+pub use iter::{UncheckedEnumerable, UncheckedIter};
 pub use map::Map;
 pub use vec::Vec;
